feat: add TikzPictureManager
